Changed i32's to usize where needed

+++ network/src/components.rs
@@ -1,16 +1,16 @@
 use innovation::NeuronType;
 
 pub struct Connector {
-    pub from: i32,
-    pub to: i32,
+    pub from: usize,
+    pub to: usize,
     pub weight: f64,
 }
 
 pub struct Neuron {
-    pub id: i32,
+    pub id: usize,
     pub kind: NeuronType,
-    pub from_arr: Vec<i32>,
-    pub to_arr: Vec<i32>,
+    pub from_arr: Vec<usize>,
+    pub to_arr: Vec<usize>,
     pub value: f64,
 }
